drop dups, move test
