+++ app/events/events_test.go
@@ -1,6 +1,14 @@
 package events
 
-import "testing"
+import (
+	"errors"
+	"testing"
+
+	tbapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/umputun/tg-spam/app/events/mocks"
+)
 
 func TestEvents_escapeMarkDownV1Text(t *testing.T) {
 	tests := []struct {
@@ -33,4 +41,44 @@ func TestEvents_escapeMarkDownV1Text(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestEvents_send(t *testing.T) {
+	mockAPI := &mocks.TbAPIMock{
+		SendFunc: func(c tbapi.Chattable) (tbapi.Message, error) {
+			if mc, ok := c.(tbapi.MessageConfig); ok {
+				if mc.Text == "badmd" && mc.ParseMode == "Markdown" {
+					return tbapi.Message{}, errors.New("bad markdown")
+				}
+			}
+			return tbapi.Message{}, nil
+		},
+	}
+
+	t.Run("send with markdown passed", func(t *testing.T) {
+		mockAPI.ResetCalls()
+		err := send(tbapi.NewMessage(123, "test"), mockAPI)
+		assert.NoError(t, err)
+		assert.Equal(t, 1, len(mockAPI.SendCalls()))
+		assert.Equal(t, int64(123), mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).ChatID)
+		assert.Equal(t, "test", mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).Text)
+		assert.Equal(t, "Markdown", mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).ParseMode)
+	})
+
+	t.Run("send with markdown failed", func(t *testing.T) {
+		mockAPI.ResetCalls()
+		err := send(tbapi.NewMessage(123, "badmd"), mockAPI)
+		assert.NoError(t, err)
+
+		assert.Equal(t, 2, len(mockAPI.SendCalls()))
+
+		assert.Equal(t, int64(123), mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).ChatID)
+		assert.Equal(t, "badmd", mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).Text)
+		assert.Equal(t, "Markdown", mockAPI.SendCalls()[0].C.(tbapi.MessageConfig).ParseMode)
+
+		assert.Equal(t, int64(123), mockAPI.SendCalls()[1].C.(tbapi.MessageConfig).ChatID)
+		assert.Equal(t, "badmd", mockAPI.SendCalls()[1].C.(tbapi.MessageConfig).Text)
+		assert.Equal(t, "", mockAPI.SendCalls()[1].C.(tbapi.MessageConfig).ParseMode)
+	})
+
 }
